+++ tests/client.py
@@ -20,12 +20,12 @@ from sora_sdk import (
     SoraSignalingDirection,
     SoraSignalingErrorCode,
     SoraSignalingType,
+    SoraTrackInterface,
     SoraVideoCodecImplementation,
     SoraVideoCodecPreference,
     SoraVideoCodecType,
     SoraVideoFrame,
     SoraVideoSink,
-    SoraVideoSource,
     get_video_codec_capability,
 )
 
@@ -68,6 +68,11 @@ class SoraClient:
         audio_output_frequency: int = 16000,
         video_width: int = 640,
         video_height: int = 480,
+        video_frame_rate: int = 30,
+        libcamera: bool = False,
+        libcamera_controls: Optional[list[tuple[str, str]]] = None,
+        native_frame_output: bool = False,
+        force_i420_conversion: Optional[bool] = None,
     ):
         self._signaling_urls = settings.signaling_urls
         self._role = role.value
@@ -105,12 +110,17 @@ class SoraClient:
 
         self._video_width: int = video_width
         self._video_height: int = video_height
+        self._video_frame_rate: int = video_frame_rate
+
+        self._libcamera = libcamera
 
         if settings.libwebrtc_log is not None:
             sora_sdk.enable_libwebrtc_log(settings.libwebrtc_log)
 
         self._sora: Sora = Sora(
-            openh264=settings.openh264_path, video_codec_preference=video_codec_preference
+            openh264=settings.openh264_path,
+            video_codec_preference=video_codec_preference,
+            force_i420_conversion=force_i420_conversion,
         )
 
         self._fake_audio_thread: Optional[threading.Thread] = None
@@ -122,8 +132,16 @@ class SoraClient:
                 self._audio_channels, self._audio_sample_rate
             )
 
-        self._video_source: Optional[SoraVideoSource] = None
-        if self._video:
+        self._video_source: Optional[SoraTrackInterface] = None
+        if libcamera and self._video:
+            self._video_source = self._sora.create_libcamera_source(
+                width=self._video_width,
+                height=self._video_height,
+                fps=self._video_frame_rate,
+                native_frame_output=native_frame_output,
+                controls=libcamera_controls,
+            )
+        elif self._video:
             self._video_source = self._sora.create_video_source()
 
         self._audio_sink: Optional[SoraAudioSink] = None
@@ -458,6 +476,9 @@ class SoraClient:
 
         self._connected.clear()
         self._disconnected.set()
+
+        if self._libcamera or self._video_source is not None:
+            self._video_source = None
 
         if self._fake_audio_thread is not None:
             self._fake_audio_thread.join(timeout=10)